+++ noodles-csi/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+  * async: Add async reader (`csi::AsyncReader`).
+
+    This can be enabled with the `async` feature.
+
 ## 0.2.2 - 2021-08-11
 
 ### Fixed